+++ eras/shelley-ma/impl/src/Cardano/Ledger/Allegra/UTxO.hs
@@ -1,0 +1,15 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Cardano.Ledger.Allegra.UTxO () where
+
+import Cardano.Ledger.Crypto
+import Cardano.Ledger.Shelley.PParams (ShelleyPParamsHKD (..))
+import Cardano.Ledger.Shelley.UTxO (EraUTxO (..), getConsumedCoin)
+import Cardano.Ledger.ShelleyMA.Era (MaryOrAllegra (Allegra), ShelleyMAEra)
+import Cardano.Ledger.ShelleyMA.TxBody ()
+
+instance Crypto c => EraUTxO (ShelleyMAEra 'Allegra c) where
+  getConsumedValue = getConsumedCoin